\title{
\packageTitle{LSPFP}
}
\name{LSPFP-internal}
\alias{bdown}
\alias{create_countintensematrix}
\alias{createFeatureTable}
\alias{createNewExp}
\alias{detectTerminal}
\alias{fasta2rds}
\alias{localfasta2rds}
\alias{getAttributeField}
\alias{getExperimentSourceFiles}
\alias{getPositions}
\alias{getTopofeatures}
\alias{gff2rds}
\alias{localgff2rds}
\alias{gffRead}
\alias{loadData}
\alias{LoadExp}
\alias{mylogisfunc}
\alias{openCSV}
\alias{openProgress}
\alias{openTXT}
\alias{peptide_filter}
\alias{print_peptides}
\alias{print_topoareas}
\alias{secscoring}
\alias{setDirStruc}
\alias{lspfp_getFastaInformation}
\alias{ss_openCSV}
\alias{ss_openTXT}
\alias{truncscoring}
\alias{wrapperPLM}
\alias{wrapperSecScore}
\description{
 Internal LSPFP functions.
}
\details{
These are not to be called by the user (or in some cases are just
waiting for proper documentation to be written :).
}
\keyword{ internal }
