\name{LSPFP-package}
\alias{LSPFP-package}
\alias{LSPFP}
\docType{package}
\title{
\packageTitle{LSPFP}
}
\description{
\packageDescription{LSPFP}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LSPFP}
\packageIndices{LSPFP}

The package offers two functions. 
Please have a look at wrapperLSPFP and printSelectedPeptides for details.
}
\author{
\packageAuthor{LSPFP}

Maintainer: \packageMaintainer{LSPFP}
}

%%\references{
%% ~~ Literature or other references for background information ~~
%%}
%% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }

%%\seealso{
%% ~~ Optional links to other man pages, e.g. ~~
%% ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%%}

\examples{
#Please have a look at wrapperLSPFP for an example.
%%~~ simple examples of the most important functions ~~
}
